+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.2 (unreleased)
 
 - Added support for text categorization
+- Added `saveToDisk` method to `NER`
 
 ## 0.1.1 (2022-08-28)